requirements.txt
fastapi
uvicorn
numpy
opencv-python
scikit-learn
pillow
python-multipart
